requirements.txt
arviz==0.23.4
fastapi==0.129.0
numpy==2.4.2
pandas==3.0.0
pymc==5.27.1
python-dotenv==1.2.1
Requests==2.32.5
scipy==1.17.0
SQLAlchemy==2.0.46
torch==2.11.0.dev20260116+cu128
transformers==4.57.2
yfinance==0.2.66

uvicorn[standard]

azure-storage-blob